+++ src/Components/Modal/EditHiringProgressModal/index.tsx
@@ -6,7 +6,7 @@ import { getValueByKey } from '../../../Utils/useGetPropertyKey';
 import selectIcon from '../../../Assets/SVG/selectIcon.svg';
 import dndIcon from '../../../Assets/SVG/dndIcon.svg';
 import * as _ from './style';
-import { DragDropContext, Draggable, Droppable } from 'react-beautiful-dnd';
+import { DragDropContext, Draggable, Droppable } from '@hello-pangea/dnd';
 import { Stack } from '@team-return/design-system';
 
 interface PropsType {
